+++ lib/screens/badges.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '/globals.dart';
 import '/utility/responsive.dart';
-import '/utility/confetti.dart';
 import 'package:confetti/confetti.dart';
 import 'dart:math';